manutenção no backend

+++ src/backend/Kairos.Presentation/Source/Features/Presenca/PresencaController.cs
@@ -2,15 +2,30 @@ namespace Kairos.Presentation.Source.Features.Presenca;
 [ApiController]
 [Route("v1/")]
 [Authorize]
-public class PresencaController(IPresencaService service) : ControllerBase
+public class PresencaController(IPresencaService service, IUsuarioService usuario) : ControllerBase
 {
     #region ListPresenca
         [HttpGet("ListPresenca")]
         [EndpointSummary("Listar todas as presenças.")]
         public async Task<ActionResult> ListPresenca([FromQuery] GetPresencaCommand command,CancellationToken token)
         {
-            var response = await service.GetHandler(command,token);
-            return Ok(response);
+            #region Authorize
+                if(User.FindFirst("id") == null)
+                {
+                    return Unauthorized("Você não está autenticado no sistema.");
+                }
+                var userId = User.GetId();
+                var user = await usuario.GetByIdHandler(new GetUsuarioByIdCommand { Id = userId }, token);
+                if(!(user.Data?.PerfilID == PerfilConstant.Adm || user.Data?.PerfilID == PerfilConstant.Organizador))
+                {
+                    return Unauthorized("Você não tem permissão para Visualizar a Dashboard.");
+                }
+            #endregion
+            
+            #region ListPresenca
+                var response = await service.GetHandler(command,token);
+                return Ok(response);
+            #endregion
         }
     #endregion
 
@@ -19,8 +34,23 @@ public class PresencaController(IPresencaService service) : ControllerBase
         [EndpointSummary("Obter presença pelo ID.")]
         public async Task<ActionResult> GetByIdPresenca([FromQuery] GetPresencaByIdCommand command, CancellationToken token)
         {
-            var response = await service.GetByIdHandler(command,token);
-            return Ok(response);
+            #region Authorize
+                if(User.FindFirst("id") == null)
+                {
+                    return Unauthorized("Você não está autenticado no sistema.");
+                }
+                var userId = User.GetId();
+                var user = await usuario.GetByIdHandler(new GetUsuarioByIdCommand { Id = userId }, token);
+                if(!(user.Data?.PerfilID == PerfilConstant.Adm || user.Data?.PerfilID == PerfilConstant.Organizador))
+                {
+                    return Unauthorized("Você não tem permissão para Visualizar a Dashboard.");
+                }
+            #endregion
+
+            #region GetByIdPresenca
+                var response = await service.GetByIdHandler(command,token);
+                return Ok(response);
+            #endregion
         }
     #endregion
 
@@ -29,8 +59,23 @@ public class PresencaController(IPresencaService service) : ControllerBase
         [EndpointSummary("Registrar uma nova presença.")]
         public async Task<IActionResult> CreatePresenca(CreatePresencaCommand command, CancellationToken token)
         {
-            var response = await service.CreateHandler(command,token);
-            return Ok(response);
+            #region Authorize
+                if(User.FindFirst("id") == null)
+                {
+                    return Unauthorized("Você não está autenticado no sistema.");
+                }
+                var userId = User.GetId();
+                var user = await usuario.GetByIdHandler(new GetUsuarioByIdCommand { Id = userId }, token);
+                if(!(user.Data?.PerfilID == PerfilConstant.Adm || user.Data?.PerfilID == PerfilConstant.Organizador))
+                {
+                    return Unauthorized("Você não tem permissão para Visualizar a Dashboard.");
+                }
+            #endregion
+            
+            #region CreatePresenca
+                var response = await service.CreateHandler(command,token);
+                return Ok(response);
+            #endregion
         }
     #endregion
 
@@ -39,8 +84,23 @@ public class PresencaController(IPresencaService service) : ControllerBase
         [EndpointSummary("Excluir presença pelo ID.")]
         public async Task<ActionResult> DeletePresenca([FromQuery] DeletePresencaCommand command, CancellationToken token)
         {
-            var response = await service.DeleteHandler(command,token);
-            return Ok(response);
+            #region Authorize
+                if(User.FindFirst("id") == null)
+                {
+                    return Unauthorized("Você não está autenticado no sistema.");
+                }
+                var userId = User.GetId();
+                var user = await usuario.GetByIdHandler(new GetUsuarioByIdCommand { Id = userId }, token);
+                if(!(user.Data?.PerfilID == PerfilConstant.Adm || user.Data?.PerfilID == PerfilConstant.Organizador))
+                {
+                    return Unauthorized("Você não tem permissão para Visualizar a Dashboard.");
+                }
+            #endregion
+            
+            #region DeletePresenca
+                var response = await service.DeleteHandler(command,token);
+                return Ok(response);
+            #endregion
         }
     #endregion
 }
